refactor: RestaurantList component 커스텀 훅 분리
